mod: add test case for jenkins;

+++ Jenkinsfile
@@ -7,7 +7,7 @@ node {
     def globalVersion = "external-oidc-provider:b${version}"
     def githubUserName = "ck-jenkins"
     def githubUserToken = "${params.githubUserToken}"
-    def userEmail = "taegeon_woo@tmax.co.kr"
+    def userEmail = "seongmin_lee2@tmax.co.kr"
 
     stage('git pull from external-oidc-provider') {
     	git branch: "${params.buildBranch}",
@@ -25,29 +25,36 @@ node {
     }
 
     stage('Dockerfile image build & push'){
-        sh "sudo docker build --tag tmaxcloudck/external-oidc-provider:${imageTag} ."
-        sh "sudo docker tag tmaxcloudck/external-oidc-provider:${imageTag} tmaxcloudck/external-oidc-provider:latest"
-        sh "sudo docker push tmaxcloudck/external-oidc-provider:${imageTag}"
-        sh "sudo docker push tmaxcloudck/external-oidc-provider:latest"
-        sh "sudo docker rmi tmaxcloudck/external-oidc-provider:${imageTag}"
+        if(type == 'distribution') {
+           sh "sudo docker build --tag tmaxcloudck/external-oidc-provider:${imageTag} ."
+           sh "sudo docker tag tmaxcloudck/external-oidc-provider:${imageTag} tmaxcloudck/external-oidc-provider:latest"
+           sh "sudo docker push tmaxcloudck/external-oidc-provider:${imageTag}"
+           sh "sudo docker push tmaxcloudck/external-oidc-provider:latest"
+           sh "sudo docker rmi tmaxcloudck/external-oidc-provider:${imageTag}"
+        } else if(type == 'test'){
+                sh "sudo docker build --tag ${imageRegistry}/external-oidc-provider:b${testVersion} ."
+                sh "sudo docker push ${imageRegistry}/external-oidc-provider:b${testVersion}"
+                sh "sudo docker rmi ${imageRegistry}/external-oidc-provider:b${testVersion}"
+        }
     }
 
     stage('git push'){
-        sh "git checkout ${params.buildBranch}"
-        sh "git add -A"
+        if(type == 'distribution') {
+            sh "git checkout ${params.buildBranch}"
+            sh "git add -A"
 
-        sh (script:'git commit -m "[Distribution] External-Oidc-Provider- ${version} " || true')
-        sh "git tag v${version}"
+            sh (script:'git commit -m "[Distribution] External-Oidc-Provider- ${version} " || true')
+            sh "git tag v${version}"
 
-        sh "git remote set-url origin https://${githubUserToken}@github.com/tmax-cloud/external-oidc-provider.git"
+            sh "git remote set-url origin https://${githubUserToken}@github.com/tmax-cloud/external-oidc-provider.git"
 
-        sh "sudo git push -u origin +${params.buildBranch}"
-        sh "sudo git push origin v${version}"
+            sh "sudo git push -u origin +${params.buildBranch}"
+            sh "sudo git push origin v${version}"
 
-        sh "git fetch --all"
-        sh "git reset --hard origin/${params.buildBranch}"
-        sh "git pull origin ${params.buildBranch}"
-
+            sh "git fetch --all"
+            sh "git reset --hard origin/${params.buildBranch}"
+            sh "git pull origin ${params.buildBranch}"
+        }
     }
 
 	stage('clear repo'){
